requirements.txt
gensim==3.4.0
nltk==3.3
summa==1.0.0
newsapi-python==0.2.2
newspaper3k==0.2.6
